New Function : 캐릭터 스킬 구현
보조 스킬 1 : 파워업
보조 스킬 2 : 스피드업 구현 완료
스킬 동시 적용 완료
스킬 통합 관리를 위한 스킬 스크립트 생성

+++ Project_A/Assets/Scripts/characterScripts/Bullet.cs
@@ -4,8 +4,10 @@ using UnityEngine;
 
 public class Bullet : MonoBehaviour
 {
-    public int damage; // 총알의 공격력
+    public float damage; // 총알의 공격력
     public bool isMelee; // 근접 공격인지 여부를 나타내는 플래그
+
+    
 
     // 충돌 시 호출되는 메서드
     void OnCollisionEnter(Collision collision)
